Module modio::auth

Authentication Flow interface

Structs

Auth
Authentication Flow interface to retrieve access tokens. See the mod.io Authentication docs for more information.

AuthOptions
Options for external authentication.

Credentials
mod.io credentials. API key with optional OAuth2 access token.

DiscordOptions
Authentication options for an Discord token.

GalaxyOptions
Authentication options for an encrypted gog app ticket.

GoogleOptions
Authentication options for an Google token.

ItchioOptions
Authentication options for an itch.io JWT token.

Link
Part of Terms

Links
Part of Terms

OculusOptions
Authentication options for an Oculus user.

SteamOptions
Authentication options for an encrypted steam app ticket.

SwitchOptions
Authentication options for the NSA ID token.

Terms
See the Terms Object docs for more information.

Token
Access token and optional Unix timestamp of the date this token will expire.

XboxOptions
Authentication options for an Xbox Live token.

Enums

Error
Authentication error

Service
The 3rd party authentication service that will be used after the user agrees to the terms of use and consent to an account being created.