Module modio::client

Re-exports of the used reqwest types.

Modules

header
HTTP header types

Structs

Certificate
Represents a server X509 certificate.

ClientBuilder
A ClientBuilder can be used to create a Client with custom configuration.

Identity
Represents a private key and X509 cert as a client certificate.

Policy
A type that controls the policy on how to handle the following of redirects.

Proxy
Configuration of a proxy that a Client should pass requests to.

Url
A parsed URL record.