Module modio::comments::filters

Comment filters and sorting.

Filters

Fulltext

Id

ModId

SubmittedBy

DateAdded

ReplyId

ThreadPosition

Karma

Content

Sorting

Id

ModId

SubmittedBy

DateAdded

See modio docs for more information.

By default this returns up to 100 items. You can limit the result by using limit and offset.

Example

use modio::filter::prelude::*;
use modio::comments::filters::Id;

let filter = Id::_in(vec![1, 2]).order_by(Id::desc());

Structs

Content

DateAdded

Fulltext

Id

Karma

ModId

ReplyId

SubmittedBy
Unique id of the user who has ownership of the objects.

ThreadPosition