Module modio::filter

Filtering and sorting

Modules

prelude
A prelude for using common filters and importing traits.

Structs

Filter

Enums

OneOrMany
Represents a value or a list of values of a filter.

Operator
Filter operators of mod.io.

Traits

BitwiseAnd

Cmp

Eq

In

Like

NotEq

NotIn

NotLike

OrderBy

Functions

custom_filter
Create a custom Filter.

custom_order_by_asc
Create a custom sorting Filter in ascending order.

custom_order_by_desc
Create a custom sorting Filter in descending order.