Module modio::filter::prelude

A prelude for using common filters and importing traits.

use modio::filter::prelude::*;

Re-exports

pub use super::BitwiseAnd;

pub use super::Cmp;

pub use super::OrderBy;

pub use super::Eq;

pub use super::NotEq;

pub use super::In;

pub use super::NotIn;

pub use super::Like;

pub use super::NotLike;

pub use super::Filter;

pub use super::OneOrMany;

Structs

DateAdded

DateLive

DateUpdated

Fulltext

Id

ModId

Name

NameId

Status

SubmittedBy
Unique id of the user who has ownership of the objects.

Functions

with_limit
Create a Filter with a limit to paginate through results.

with_offset
Create a Filter with an offset to paginate through results.