Module modio::games

Games interface

Modules

filters
Game filters and sorting.

Structs

AddTagsOptions

ApiAccessOptions
Level of API access allowed by a game.

CommunityOptions
Community features enabled on the mod.io website.

DeleteTagsOptions

Downloads
Part of Statistics

EditGameOptions

EditMediaOptions

Game
See the Game Object docs for more information.

GameRef
Reference interface of a game.

Games
Interface for games.

HeaderImage
See the Header Image Object docs for more information.

Icon
See the Icon Object docs for more information.

Logo
See the Logo Object docs for more information.

RevenueOptions
Revenue capabilities mods can enable.

Statistics
See the Game Statistics Object docs for more information.

TagOption
See the Game Tag Option Object docs for more information.

Tags
Interface for tag options.

Enums

CurationOption
Curation process used to approve mods.

MaturityOptions
Option to allow developers to select if they flag their mods containing mature content.

PresentationOption
Presentation style used on the mod.io website.

Status
See Status & Visibility docs for more information.

SubmissionOption
Submission process modders must follow.

TagType
Defines the type of a tag. See TagOption.