Module modio::mods

Mods Interface

Modules

filters
Mod filters & sorting

Structs

AddMediaOptions

AddModOptions

DeleteMediaOptions

Dependencies
Interface for dependencies.

Dependency
See the Mod Dependency Object docs for more information.

EditDependenciesOptions

EditModOptions

EditTagsOptions

Event
See the Mod Event Object docs for more information.

Image
See the Image Object docs for more information.

Logo
See the Logo Object docs for more information.

MaturityOption
Maturity options a mod can be flagged.

Media
See the Mod Media Object docs for more information.

Mod
See the Mod Object docs for more information.

ModRef
Reference interface of a mod.

Mods
Interface for mods of a game.

Popularity
Part of Statistics

Ratings
Part of Statistics

Statistics
See the Statistics Object docs for more information.

Tag
See the Mod Tag Object docs for more information.

Tags
Interface for tags.

Enums

EventType
Type of mod event that was triggered.

Rating

Status
See Status & Visibility docs for more information.

Visibility
See Status & Visibility docs for more information.