Struct modio::mods::ModRef

pub struct ModRef<C> where
    C: Clone + Connect + 'static,  { /* fields omitted */ }

Reference interface of a mod.

Methods

impl<C: Clone + Connect + 'static> ModRef<C>

pub fn get(&self) -> Future<Mod>

Get a reference to the Modio mod object that this ModRef refers to.

pub fn files(&self) -> Files<C>

Return a reference to an interface that provides access to the files of a mod.

pub fn file(&self, id: u32) -> FileRef<C>

Return a reference to a file of a mod.

pub fn metadata(&self) -> Metadata<C>

Return a reference to an interface to manage metadata key value pairs of a mod.

pub fn tags(&self) -> Endpoint<C, Tag>

Return a reference to an interface to manage the tags of a mod.

pub fn comments(&self) -> Comments<C>

Return a reference to an interface that provides access to the comments of a mod.

pub fn dependencies(&self) -> Endpoint<C, Dependency>

Return a reference to an interface to manage the dependencies of a mod.

pub fn events(
&self,
options: &EventListOptions
) -> Future<ModioListResponse<Event>>

Return the event log for a mod sorted by latest event first.

pub fn members(&self) -> Members<C>

Return a reference to an interface to manage team members of a mod.

pub fn edit(&self, options: &EditModOptions) -> Future<Mod>

Edit details for a mod.

pub fn add_media(&self, options: AddMediaOptions) -> Future<ModioMessage>

Add new media to a mod.

pub fn delete_media(&self, options: &DeleteMediaOptions) -> Future<()>

Delete media from a mod.

pub fn rate(&self, rating: Rating) -> Future<()>

Submit a positive or negative rating for a mod.

pub fn subscribe(&self) -> Future<()>

Subscribe the authenticated user to a mod.

pub fn unsubscribe(&self) -> Future<()>

Unsubscribe the authenticated user from a mod.