Module modio::user

User interface

Modules

filters
Filters for events, subscriptions and ratings.

Structs

Avatar
See the Avatar Object docs for more information.

Event
See the User Event Object docs for more information.

Me
Interface for resources owned by the authenticated user or is team member of.

User
See the User Object docs for more information.

Enums

EventType
Type of user event that was triggered.

Rating
See the Rating Object docs for more information.