Module modio::user::filters::games

Game filters and sorting.

Filters

Fulltext

Id

Status

SubmittedBy

DateAdded

DateUpdated

DateLive

Name

NameId

Summary

InstructionsUrl

UgcName

PresentationOption

SubmissionOption

CurationOption

CommunityOptions

RevenueOptions

ApiAccessOptions

MaturityOptions

Sorting

Id

Status

Name

NameId

DateUpdated

See modio docs for more information.

By default this returns up to 100 items. You can limit the result by using limit and offset.

Example

use modio::filter::prelude::*;
use modio::games::filters::Id;

let filter = Id::_in(vec![1, 2]).order_by(Id::desc());

Structs

ApiAccessOptions

CommunityOptions

CurationOption

DateAdded

DateLive

DateUpdated

Fulltext

Id

InstructionsUrl

MaturityOptions

Name

NameId

PresentationOption

RevenueOptions

Status

SubmissionOption

SubmittedBy
Unique id of the user who has ownership of the objects.

Summary

UgcName