Module modio::user::filters::subscriptions

Subscriptions filters and sorting.

Filters

Fulltext

Id

GameId

Status

Visible

SubmittedBy

DateAdded

DateUpdated

DateLive

MaturityOption

Name

NameId

Summary

Description

Homepage

Modfile

MetadataBlob

MetadataKVP

Tags

Sorting

Id

Name

Downloads

Popular

Ratings

Subscribers

See the mod.io docs for more information.

By default this returns up to 100 items. you can limit the result by using limit and offset.

Example

use modio::filter::prelude::*;
use modio::user::filters::subscriptions::Id;

let filter = Id::_in(vec![1, 2]).order_by(Id::desc());

Structs

DateAdded

DateLive

DateUpdated

Description

Downloads

Fulltext

GameId

Homepage

Id

MaturityOption

MetadataBlob

MetadataKVP

Modfile

Name

NameId

Popular

Ratings

Status

SubmittedBy
Unique id of the user who has ownership of the objects.

Subscribers

Summary

Tags

Visible