Module modio::games

Games interface

Structs

AddTagsOptions

DeleteTagsOptions

Game
See the Game Object docs for more informations.

GameMediaOptions

GameMediaOptionsBuilder

GameRef
Reference interface of a game.

Games
Interface for games.

GamesListOptions
Options used to filter game listings.

HeaderImage
See the Header Image Object docs for more informations.

Icon
See the Icon Object docs for more informations.

Logo
See the Logo Object docs for more informations.

MyGames
Interface for games the authenticated user added or is team member of.

TagOption
See the Game Tag Option Object docs for more informations.

Enums

TagType