Module modio::mods

Mods Interface

Structs

AddMediaOptions

AddMediaOptionsBuilder

AddModOptions

AddModOptionsBuilder

DeleteMediaOptions

DeleteMediaOptionsBuilder

Dependency
See the Mod Dependency Object docs for more information.

EditDepencenciesOptions

EditModOptions

EditModOptionsBuilder

EditTagsOptions

Event
See the Mod Event Object docs for more information.

Image
See the Image Object docs for more information.

Logo
See the Logo Object docs for more information.

Media
See the Mod Media Object docs for more information.

Mod
See the Mod Object docs for more information.

ModRef
Reference interface of a mod.

Mods
Interface for mods of a game.

ModsListOptions
Options used to filter mod listings.

MyMods
Interface for mods the authenticated user added or is team member of.

Popularity

Ratings

Statistics
See the Statistics Object docs for more information.

StatsListOptions
Options used to filter mod statistics.

Tag
See the Mod Tag Object docs for more information.

Enums

EventType

Rating

Type Definitions

MetadataMap
See the Metadata KVP Object docs for more information.