Struct modio::Modio

pub struct Modio<C> where
    C: Clone + Connect + 'static,  { /* fields omitted */ }

Endpoint interface to interacting with the mod.io API.

Methods

impl Modio<HttpsConnector<HttpConnector>>

pub fn new<A, C>(agent: A, credentials: C) -> Self where
A: Into<String>,
C: Into<Credentials>,

Create an endpoint to https://api.mod.io/v1.

pub fn host<H, A, C>(host: H, agent: A, credentials: C) -> Self where
H: Into<String>,
A: Into<String>,
C: Into<Credentials>,

Create an endpoint to a different host.

impl<C> Modio<C> where
C: Clone + Connect + 'static,

pub fn custom<H, A, CR>(
host: H,
agent: A,
credentials: CR,
client: Client<C>
) -> Self where
H: Into<String>,
A: Into<String>,
CR: Into<Credentials>,

Create an endpoint with a custom hyper client.

pub fn with_credentials<CR>(self, credentials: CR) -> Self where
CR: Into<Credentials>,

Consume the endpoint and create an endpoint with new credentials.

pub fn auth(&self) -> Auth<C>

Return a reference to an interface for requesting access tokens.

pub fn games(&self) -> Games<C>

Return a reference to an interface that provides access to game informations.

pub fn game(&self, game_id: u32) -> GameRef<C>

Return a reference to a game.

pub fn mod_(&self, game_id: u32, mod_id: u32) -> ModRef<C>

Return a reference to a mod.

pub fn me(&self) -> Me<C>

Return a reference to an interface that provides access to resources owned by the user associated with the current authentication credentials.

pub fn users(&self) -> Users<C>

Return a reference to an interface that provides access to user informations.

pub fn reports(&self) -> Reports<C>

Return a reference to an interface to report games, mods and users.